Apps: Fix layout

+++ app/src/main/java/foundation/e/apps/api/data/Ratings.kt
@@ -1,6 +1,6 @@
 package foundation.e.apps.api.data
 
 data class Ratings(
-    val privacyScore: Int,
-    val usageQualityScore: Int
+    val privacyScore: Double,
+    val usageQualityScore: Double
 )
